Refactor duplication dialog: replace DuplicationInfoDialog with RoomCopyInfoDialog and update related tests
