+++ src/modules/feature/room/RoomCopyInfoDialog.vue
